+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Released on Sunday, July 26 2026
 - Improved observer dispatch during HTML parsing (#1266) @dv00d00
 - Improved declaration of unsupported pseudo elements
 - Fixed inconsistent behavior and segfaults on Android Release build (#1252) @DevGitPit
+- Fixed CSS attribute selector modifier handling for `i` / `I` (case-insensitive) and `s` / `S` (case-sensitive)
 - Added API enhancements to expose asynchronous HTML token source (#1265) @dv00d00
 - Added bounded stream parsing mode (#1263) @dv00d00
 - Added canonicalization of decimal, octal, or hexadecimal IPv4 literals in `Url`